requirements.txt
numpy
pandas
matplotlib
configparser
google-search-results
google-cloud-bigquery
google-cloud-storage
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
torch
torchvision
transformers
pandas-gbq